+++ Realm/ObjectStore/list.hpp
@@ -91,16 +91,16 @@ public:
     // The List object has been invalidated (due to the Realm being invalidated,
     // or the containing object being deleted)
     // All non-noexcept functions can throw this
-    struct InvalidatedException {};
+    struct InvalidatedException : public std::runtime_error {
+        InvalidatedException() : std::runtime_error("Access to invalidated List object") {}
+    };
 
     // The input index parameter was out of bounds
-    struct OutOfBoundsIndexException {
+    struct OutOfBoundsIndexException : public std::out_of_range {
+        OutOfBoundsIndexException(size_t r, size_t c);
         size_t requested;
         size_t valid_count;
     };
-
-    // The input Row object is not attached
-    struct DetatchedAccessorException { };
 
 private:
     std::shared_ptr<Realm> m_realm;