+++ Realm/Tests/Swift2.2/SwiftPropertyTypeTest.swift
@@ -103,4 +103,20 @@ class SwiftPropertyTypeTest: RLMTestCase {
         XCTAssertEqual(obj.int32, v32)
         XCTAssertEqual(obj.int64, v64)
     }
+
+    func testLazyVarProperties() {
+        let realm = realmWithTestPath()
+        let succeeded : Void? = try? realm.transactionWithBlock() {
+            realm.addObject(SwiftLazyVarObject())
+        }
+        XCTAssertNotNil(succeeded, "Writing an NSObject-based object with an lazy property should work.")
+    }
+
+    func testIgnoredLazyVarProperties() {
+        let realm = realmWithTestPath()
+        let succeeded : Void? = try? realm.transactionWithBlock() {
+            realm.addObject(SwiftIgnoredLazyVarObject())
+        }
+        XCTAssertNotNil(succeeded, "Writing an object with an ignored lazy property should work.")
+    }
 }